backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
loguru
httpx
pytest
pydantic